+++ libs/cardano-ledger-api/test/Test/Cardano/Ledger/Api/State/QuerySpec.hs
@@ -82,6 +82,7 @@ latestErasSpec =
           roundTripEraExpectation @era @QueryResultCommitteeMembersState
         prop "QueryResultConstitution" $
           roundTripEraExpectation @era @QueryResultConstitution
+        prop "QueryResultDelegsAndRewards" $ roundTripEraExpectation @era @QueryResultDelegsAndRewards
         prop "QueryResultDRepState" $
           roundTripEraExpectation @era @QueryResultDRepState
         prop "QueryResultDRepStates" $